WIP: Analytics COMPILES
