+++ src/analytics/src/main/scala/osmesa/analytics/EditHistogram.scala
@@ -2,18 +2,15 @@ package osmesa.analytics
 
 import java.net.URI
 
-import geotrellis.raster.RasterExtent
 import geotrellis.layer._
-import geotrellis.store.index.zcurve.ZSpatialKeyIndex
 import geotrellis.vector._
-import geotrellis.vectortile.{MVTFeature, VInt64, Value, VectorTile}
+import geotrellis.vectortile.{MVTFeature, VInt64, VectorTile}
 import org.apache.spark.sql._
 import org.locationtech.jts.geom.util.AffineTransformation
 import osmesa.analytics.updater.Implicits._
 import osmesa.analytics.updater.{makeLayer, path, write}
 import osmesa.analytics.vectorgrid._
 
-import scala.collection.mutable.ArrayBuffer
 import scala.collection.parallel.{ForkJoinTaskSupport, TaskSupport}
 import scala.collection.{GenIterable, GenMap}
 import scala.concurrent.forkjoin.ForkJoinPool
@@ -306,85 +303,4 @@ object EditHistogram extends VectorGrid {
       }
       .filterNot(x => x.features.isEmpty)
 
-  object implicits {
-    implicit class ExtendedPointFeatureTraversableOnce(
-        val features: TraversableOnce[PointFeature[Map[String, Long]]]) {
-
-      /**
-        * Merge features that are part of the same vectorized tile and contain distinct properties.
-        *
-        * @return Merged features
-        */
-      def merge(): Seq[PointFeature[Map[String, Long]]] =
-        features
-          .foldLeft(Map.empty[Long, PointFeature[Map[String, Long]]]) {
-            case (acc, feat) =>
-              val sk = feat.data("__id")
-
-              val merged = acc.get(sk) match {
-                case Some(f) => f.mapData(d => d ++ feat.data)
-                case None    => feat
-              }
-
-              acc.updated(sk, merged)
-          }
-          .values
-          .toVector // materialize iterator
-    }
-
-    implicit class RasterTileWithKeyExtension(tiles: Dataset[RasterTileWithKey]) {
-      import tiles.sparkSession.implicits._
-
-      def vectorize: Dataset[VectorTileWithKey] =
-        tiles.map { tile =>
-          // convert into features
-          val raster = tile.raster
-          val rasterExtent =
-            RasterExtent(raster.extent, raster.tile.cols, raster.tile.rows)
-          val index = new ZSpatialKeyIndex(
-            KeyBounds(SpatialKey(0, 0), SpatialKey(raster.tile.cols - 1, raster.tile.rows - 1)))
-
-          val features = ArrayBuffer[PointFeature[Map[String, Long]]]()
-
-          raster.tile.foreach { (c: Int, r: Int, value: Int) =>
-            if (value > 0) {
-              features.append(
-                PointFeature(Point(rasterExtent.gridToMap(c, r)),
-                             Map(tile.key -> value,
-                                 "__id" -> index.toIndex(SpatialKey(c, r)).toLong)))
-            }
-          }
-
-          VectorTileWithKey(tile.key, tile.zoom, tile.sk, features)
-        }
-    }
-
-    implicit class RasterTileWithKeyAndSequenceExtension(
-        tiles: Dataset[RasterTileWithKeyAndSequence]) {
-      import tiles.sparkSession.implicits._
-
-      def vectorize: Dataset[VectorTileWithKeyAndSequence] =
-        tiles.map { tile =>
-          // convert into features
-          val raster = tile.raster
-          val rasterExtent =
-            RasterExtent(raster.extent, raster.tile.cols, raster.tile.rows)
-          val index = new ZSpatialKeyIndex(
-            KeyBounds(SpatialKey(0, 0), SpatialKey(raster.tile.cols - 1, raster.tile.rows - 1)))
-
-          val features = ArrayBuffer[PointFeature[Map[String, Long]]]()
-
-          raster.tile.foreach { (c: Int, r: Int, value: Int) =>
-            if (value > 0) {
-              features.append(
-                PointFeature(Point(rasterExtent.gridToMap(c, r)),
-                             Map(tile.key -> value,
-                                 "__id" -> index.toIndex(SpatialKey(c, r)).toLong)))
-            }
-          }
-
-          VectorTileWithKeyAndSequence(tile.sequence, tile.key, tile.zoom, tile.sk, features)
-        }
-    }
-  }
 }
