+++ src/analytics/src/main/scala/osmesa/analytics/updater/schemas/History.scala
@@ -3,7 +3,7 @@ package osmesa.analytics.updater.schemas
 import java.sql.Timestamp
 import java.time.Instant
 
-import geotrellis.vector.{Feature, Geometry}
+import geotrellis.vector.Geometry
 import geotrellis.vectortile._
 import osmesa.analytics.updater.Implicits._
 import osmesa.analytics.updater._
